+++ app/src/main/java/com/dima/financeapp/model/net/RecordResponse.kt
@@ -3,7 +3,7 @@ package com.dima.financeapp.model.net
 data class RecordResponse(
     val id: Int,
     val name: String,
-    val sum: Int,
+    val sum: Double,
     val type: String,
     val color: Int,
     val icon: Int,
